requirements.txt
PyJWT>=2.0.0
requests>=2.0.0